+++ lib/reminders.dart
@@ -8,24 +8,25 @@ export 'reminder.dart';
 class Reminders {
   static const MethodChannel _channel = const MethodChannel('reminders');
 
-  /// Check whether you have access to the ios Reminders
-  static Future<bool> get hasAccess async {
+  /// Check for access to the ios Reminders
+  Future<bool> get hasAccess async {
     return await _channel.invokeMethod('hasAccess');
   }
 
-  /// Returns the name of the default list of reminders
-  static Future<String> get defaultList async {
-    final String list = await _channel.invokeMethod('getDefaultList');
+  /// Return the name of the default list of reminders
+  Future<String> get defaultList async {
+    final String list = await _channel.invokeMethod('defaultList');
     return list;
   }
 
-  /// Returns a list of all the reminder list names
-  static Future<List<dynamic>> get allLists async {
-    return await _channel.invokeMethod("getAllLists");
+  /// Return a list of all the reminder list names
+  Future<List<dynamic>> get lists async {
+    return await _channel.invokeMethod("lists");
   }
 
-  /// Returns a list of [Reminder]s found in a list
-  static Future<List<Reminder>> getReminders(String list) async {
+  /// Return a list of [Reminder]s found in a list
+  /// Pass an empty string to return all reminders in all lists
+  Future<List<Reminder>> getReminders(String list) async {
     var reminders =
         await _channel.invokeMethod("getReminders", {"calendar": list});
     return reminders
@@ -34,7 +35,7 @@ class Reminders {
   }
 
   /// Deletes a [Reminder] given it's id
-  static Future deleteReminder(String reminder) async {
-    await _channel.invokeMethod("deleteReminder", {"reminder": reminder});
+  Future deleteReminderWithId(String reminder) async {
+    await _channel.invokeMethod("deleteReminderWithId", {"reminder": reminder});
   }
 }