+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.1.0 2020-12-16
+
+* Update README with correct Key Value pair in info.plist
+* Change Reminders API to an object from a static class
+* Tidy API method names
+
 ## 0.0.2 2020-11-30
 
 * Add API documentation